--- apps/gotask_api/src/domain/model/otp/refreshToken.ts
@@ -1,41 +1,0 @@
-import { Schema, model, Document, Types } from "mongoose";
-
-export interface IRefreshToken extends Document {
-  user: Types.ObjectId;
-  token: string;
-  expiresAt: Date;
-  isRevoked: boolean;
-  createdAt?: Date;
-  updatedAt?: Date;
-}
-
-const RefreshTokenSchema = new Schema<IRefreshToken>(
-  {
-    user: {
-      type: Schema.Types.ObjectId,
-      ref: "User",
-      required: true
-    },
-    token: {
-      type: String,
-      required: true,
-      unique: true
-    },
-    expiresAt: {
-      type: Date,
-      required: true
-    },
-    isRevoked: {
-      type: Boolean,
-      default: false
-    }
-  },
-  {
-    timestamps: true // Adds createdAt and updatedAt
-  }
-);
-
-// TTL index: MongoDB will auto-delete documents after expiresAt
-RefreshTokenSchema.index({ expiresAt: 1 }, { expireAfterSeconds: 0 });
-
-export const RefreshToken = model<IRefreshToken>("RefreshToken", RefreshTokenSchema);
